+++ test/clj/spring_lobby/http_test.clj
@@ -56,12 +56,18 @@
            (http/bar-engine-filename "104.0.1-1695-gbd6b256 BAR"))))
   (is (= "spring_bar_.BAR.104.0.1.1828-g1f481b7_windows-64-minimal-portable.7z"
          (with-redefs [fs/get-sys-data (constantly {:os-name "Windows"})]
-           (http/bar-engine-filename "104.0.1.1828-g1f481b7 BAR")))))
+           (http/bar-engine-filename "104.0.1.1828-g1f481b7 BAR"))))
+  (is (= "spring_bar_.BAR105.105.0-430-g2727993_windows-64-minimal-portable.7z"
+         (with-redefs [fs/get-sys-data (constantly {:os-name "Windows"})]
+           (http/bar-engine-filename "105.0-430-g2727993 BAR105")))))
 
 (deftest bar-engine-filename?
   (is (false?
         (http/bar-engine-filename?
-          "spring_bar_.BAR.104.0.1-1656-gad7994a_windows-64-minimal-symbols.tgz",)))
+          "spring_bar_.BAR.104.0.1-1656-gad7994a_windows-64-minimal-symbols.tgz")))
   (is (true?
         (http/bar-engine-filename?
-          "spring_bar_.BAR.104.0.1-1656-gad7994a_windows-64-minimal-portable.7z",))))
+          "spring_bar_.BAR.104.0.1-1656-gad7994a_windows-64-minimal-portable.7z")))
+  (is (true?
+        (http/bar-engine-filename?
+          "spring_bar_.BAR105.105.0-430-g2727993_windows-64-minimal-portable.7z"))))